ATmega32: More work on separating out USART to compose off of 328p
